JTAG: async OnXxxShift via JtagPending<T>; retire SPI/TIM variants
Every ITapInterface::OnXxxShift() now returns a JtagPending<T> (POD: 8 B,
rx pointer + decoder fn). Each shift renders frame N+1 in parallel with
frame N's still-running DMA, then waits on N at the top of the new shift.
Implicit operator T() resolves on demand; statement-only calls are
fire-and-forget — TapPlayer's variadic batch loop pipelines for free.

Drains added in AcquireSpiClkMuted / AcquireGpioMode / OnFlashTclk /
OnReleaseJtag / OnResetTap / OnClose so non-shift SPI access can't race
an in-flight shift DMA.

Removes the SPI (OPT_JTAG_IMPL_SPI / SPI_DMA) and TIM_DMA
(OPT_JTAG_IMPL_TIM_DMA / TIM_DMA_SLOW) variants — DTRIG covers every
supported target. See .claude/docs/drivers/SPI_VARIANT_REMOVED.md and
TIM_VARIANT_REMOVED.md for last-working git refs and rationale. Nucleo
L432KC + G431KB platform.h ported to DTRIG (untested — flagged in-file).

+++ Firmware.shared/drivers/JtagDev.cpp
@@ -141,7 +141,7 @@ void JtagDev::OnEnterTap()
 bool JtagDev::IsInstrLoad()
 {
 	OnIrShift(IR_CNTRL_SIG_CAPTURE);
-	CtrlSigReg reg = static_cast<CtrlSigReg>(OnDrShift16(0));
+	CtrlSigReg reg = static_cast<CtrlSigReg>((uint16_t)OnDrShift16(0));
 	constexpr CtrlSigReg mask = CtrlSigReg::kRead | CtrlSigReg::kInstrLoad;
 	return IsSet(reg, mask, mask);
 }
